requirements-v5.txt
# Omnipath v5.0 Additional Dependencies
# Add these to your existing requirements.txt

# NATS Event Bus
nats-py==2.6.0

# OpenTelemetry (Observability)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# CLI Dependencies (optional - only if using CLI)
typer[all]==0.9.0
rich==13.7.0
httpx==0.25.2
